src/common/evaluations/pyproject.toml
[project]
name = "evaluations"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11,<3.12"
dependencies = [
    "flask>=3.1.0",
    "gunicorn>=23.0.0",
    "mysql-connector-python>=9.4.0",
    "psutil>=7.0.0",
    "pyyaml>=6.0.2",
    "requests>=2.32.5",
    "werkzeug>=2.3.8",
]

[project.optional-dependencies]
selenium = [
    "selenium>=4.35.0",
]
